ARCH-1: extract okf/madr dispatch arms (iter-225)
