+++ crates/hyalo-cli/src/commands/okf.rs
@@ -1791,3 +1791,54 @@ mod tests {
         assert_eq!(indent_continuation("a\r\nb"), "a\n  b");
     }
 }
+
+// ---------------------------------------------------------------------------
+// Dispatch handler (ARCH-1, iter-225)
+// ---------------------------------------------------------------------------
+
+/// The `hyalo okf` dispatch arm, extracted verbatim from `dispatch.rs`.
+pub(crate) fn run(
+    ctx: &mut crate::dispatch::CommandContext<'_>,
+    action: crate::cli::args::OkfAction,
+) -> Result<CommandOutcome> {
+    let effective_format = ctx.effective_format;
+    use hyalo_core::case_index::mode_enabled;
+
+match action {
+    crate::cli::args::OkfAction::Index {
+        scope,
+        apply,
+        dry_run: _,
+        replace,
+    } => {
+        let case_insensitive = mode_enabled(ctx.case_insensitive_mode, ctx.dir);
+        let (outcome, exit_override) = crate::commands::okf::run_index(
+            ctx.dir,
+            scope.as_deref(),
+            apply,
+            replace,
+            ctx.okf_ignore,
+            case_insensitive,
+            effective_format,
+        )?;
+        if let Some(code) = exit_override {
+            ctx.exit_code_override = Some(code);
+        }
+        Ok(outcome)
+    }
+    crate::cli::args::OkfAction::Log {
+        target,
+        message,
+        action: log_action,
+        apply,
+        dry_run: _,
+    } => crate::commands::okf::run_log(
+        ctx.dir,
+        target.as_deref(),
+        &message,
+        log_action.as_deref(),
+        apply,
+        effective_format,
+    ),
+    }
+}
